apps/kyc/requirements.txt
# requirements must be compatible with python 3.13.3
easyocr==1.7.0
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.0.74
